Player: own playlist index selection and shuffle bag in navigation policy
Prior green actual-host origin pin: fbed7e8f96bbba502bd8832345a63d84ce9bbe42
on source 4d695e52cb5d5ad058f2c16b296f3b09dd81be0e; five cases green before
this move. The corrected pin preserves first analyzer INFO and first runtime
4PASS/1ERROR chronology. Test bytes remain a70da07d unchanged.

Apply independently source-accepted two-path candidate
19c996ee95fadfb9a53823ffd7b5ea16ab62c2c55015764209e4cdb4134a44d7.
Five body inverse maps are in
ignored navigation-prep/candidate, with full caller/state ledger. Retain actual
40% threshold (including stale 70% comment), unknown-size/year comparator,
series missing-index/catch asymmetry, bag filtering/refill/removeLast order,
and the same borrowed Random. Three paired writes remain in their original
setState closures; clear-only remains outside. No loading/source/state proxy.

Next/previous preserve lazy getter then conditional config short-circuit.
Shuffle samples the concrete final widget enum after empty guard/lazy getter
but before synchronous series eligibility return: reviewed behavioral read-order
equivalence, not literal/verbatim order. Parameter promotion removes field !.

Host 10895->10720 (-175 physical), owner229, whole production+54. Six operations
and a read-only flag; two owned cells, no callbacks. Four host adapters/28 lines
remain: expiry Phase3 caller ownership; this is not completed navigation or
whole-method retirement. No native/all-branch/random-sequence proof.

Verification evidence and exact commands: ignored navigation-prep/product-validation.
One production batch20PASS/0fail/0skip, raw4df15d4ab2b78b07526a1b2957dde3b97ef4f5844209a3c886a4a8e6751f7907.
Paired lib/test analyzer430->430 against449baseline, zero new identities;
touched58 inherited host diagnostics, owner/pin0. Same checker layering53->53,
zero new IDs. Initial ignored raw-byte verifier STOP retained: git apply
core.autocrlf produced CRLF; exact LF payloads proved unchanged, only unrun
checks resumed. No analyzer/test retry or production repair.
No baseline, test, configuration, CODEMAP, SDK or other production changes.
